moved _parse_species_list function to util.py
