+++ .mailmap
@@ -283,3 +283,5 @@ Israel Roldan <israel.alberto.rv@gmail.com> AirvZxf <israel.alberto.rv@gmail.com
 Israel Roldan <israel.alberto.rv@gmail.com> airv_zxf <israel.alberto.rv@gmail.com>
 
 Michael Zingale <michael.zingale@stonybrook.edu>
+
+Rudraksh Nalbalwar <nalbalwarrudraksh@gmail.com>
